+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - agent:
   - Record slurmrestd cache hit and miss counts in cache service.
   - Add metrics with cache hit and miss counts.
+  - Support querying cache metrics.
 - conf:
   - Introduce `view-cache` authorization action.
   - Assign `view-cache` action to _user_ role (all authenticated users) in